+++ deploy/complete/terraform/mushop-variables.tf
@@ -4,14 +4,67 @@
 
 # MuShop
 ## Ingress/LoadBalancer
+variable "ingress_nginx_enabled" {
+  default     = true
+  description = "Enable Ingress Nginx for Kubernetes Services (This option provision a Load Balancer)"
+}
 variable "ingress_load_balancer_shape" {
-  default = "100Mbps"
+  default     = "flexible" # Flexible, 10Mbps, 100Mbps, 400Mbps or 8000Mps
+  description = "Shape that will be included on the Ingress annotation for the OCI Load Balancer creation"
+}
+variable "ingress_load_balancer_shape_flex_min" {
+  default     = "10"
+  description = "Enter the minimum size of the flexible shape."
+}
+variable "ingress_load_balancer_shape_flex_max" {
+  default     = "100"
+  description = "Enter the maximum size of the flexible shape (Should be bigger than minimum size). The maximum service limit is set by your tenancy limits."
+}
+variable "ingress_hosts" {
+  default     = ""
+  description = "Enter a valid full qualified domain name (FQDN). You will need to map the domain name to the EXTERNAL-IP address on your DNS provider (DNS Registry type - A). If you have multiple domain names, include separated by comma. e.g.: mushop.example.com,catshop.com"
+}
+variable "cert_manager_enabled" {
+  default     = true
+  description = "Enable x509 Certificate Management"
+}
+variable "ingress_tls" {
+  default     = false
+  description = "If enabled, will generate SSL certificates to enable HTTPS for the ingress using the Certificate Issuer"
+}
+variable "ingress_cluster_issuer" {
+  default     = "letsencrypt-prod"
+  description = "Certificate issuer type. Currently supports the free Let's Encrypt and Self-Signed. Only *letsencrypt-prod* generates valid certificates"
+}
+variable "ingress_email_issuer" {
+  default     = "no-reply@mushop.ateam.cloud"
+  description = "You must replace this email address with your own. The certificate provider will use this to contact you about expiring certificates, and issues related to your account."
 }
 
 ## Features
 variable "mushop_mock_mode_all" {
   default     = false
   description = "If enabled, will not provision ATP Database, Object Storage, or Streaming. The services will run in mock mode and will connect to an in-memory database, and the data will not persist"
+}
+variable "create_oci_service_user" {
+  default     = true
+  description = "Creates OCI Service User. Service user is needed for Email Delivery (Newsletter feature) and Stream services."
+}
+variable "newsletter_subscription_enabled" {
+  default     = true
+  description = "Enables newsletter subscription feature. Deploys API Gateway, Newsletter Function and uses Email Sender service"
+}
+variable "newsletter_email_sender" {
+  default     = "no-reply@mushop.ateam.cloud"
+  description = "Email sender suffix for Newsletter Subscription. The deployment id will added as email suffix. e.g.: no-reply+xyz1@mushop.ateam.cloud"
+}
+variable "newsletter_subscription_function_image" {
+  default     = "iad.ocir.io/ociateam/mushop/newsletter-subscription"
+  description = "Container Image for the Newsletter Subscription Function"
+}
+variable "newsletter_subscription_function_image_version" {
+  default     = "0.1.0"
+  description = "Container Image Version for the Newsletter Subscription Function"
 }
 
 ## Common Services (MuShop Utilities)
@@ -30,14 +83,6 @@ variable "metrics_server_enabled" {
 variable "catalog_enabled" {
   default     = false
   description = "Enable Service Catalog to use with OCI Service Broker"
-}
-variable "ingress_nginx_enabled" {
-  default     = true
-  description = "Enable Ingress Nginx for Kubernetes Services (This option provision a Load Balancer)"
-}
-variable "cert_manager_enabled" {
-  default     = true
-  description = "Enable x509 Certificate Management"
 }
 
 ## Secrets