+++ deploy/complete/terraform/oke-variables.tf
@@ -147,6 +147,7 @@ variable "network_cidrs" {
     VCN-CIDR                      = "10.20.0.0/16"
     SUBNET-REGIONAL-CIDR          = "10.20.10.0/24"
     LB-SUBNET-REGIONAL-CIDR       = "10.20.20.0/24"
+    APIGW-FN-SUBNET-REGIONAL-CIDR = "10.20.30.0/24"
     ENDPOINT-SUBNET-REGIONAL-CIDR = "10.20.0.0/28"
     ALL-CIDR                      = "0.0.0.0/0"
     PODS-CIDR                     = "10.244.0.0/16"