+++ deploy/complete/terraform/CHANGELOG.md
@@ -1,5 +1,33 @@
 # CHANGELOG
 
+2021-06-09 (v2.4.0)
+
+- E-mail newsletter service included on the stack
+- Leveraging and automatically deploying OCI API Gateway (newsletter)
+- Leveraging and automatically deploying OCI Functions (newsletter)
+- Leveraging OCI Email Delivery (newsletter)
+- Multi-arch support for Catalogue Service (amd64 and Arm64)
+- Multi-arch support for Payment Service (amd64 and Arm64)
+- Multi-arch support for Assets Service (amd64 and Arm64)
+- Multi-arch support for Storefront Service (amd64 and Arm64)
+- Multi-arch support for API Service (amd64 and Arm64)
+- Multi-arch support for dbtools (amd64 and Arm64)
+- Policies improved
+- Cluster Utilities versions updated
+- Multiple small improvements and fixes
+- DB scripts improvements
+- catalogue service tweaks
+
+2021-05-18 (v2.3.0)
+
+- Support for flexible Load Balancer annotations on Ingress-Nginx
+- Support for domain name (FQDN) for ingress
+- Better variables and schema for Certificate management
+- Cert-manager updates
+- Schema updates for Ingress new features
+- Cluster Utilities version updates
+- apiVersion and manifest updates on deprecated items
+
 2021-04-08 (v2.2.0)
 
 - OKE Cluster Autoscaler support